+++ CHANGELOG.md
@@ -9,6 +9,19 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ...
 
+## [2.0.0] - 2019-03-11
+
+This is a **breaking release** that upgrades to the new asynchronous interfaces
+in `[mvxcvi/blocks 2.0.0]`.
+
+### Added
+- Block store accepts a `:credentials-provider` which will be used if an AWS
+  access key and secret are not provided directly.
+
+### Changed
+- Upgrade Clojure to 1.10.0
+- Upgrade to new async storage interfaces.
+
 ## [0.4.0] - 2018-03-22
 
 ### Changed
@@ -70,7 +83,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 Initial library release.
 
-[Unreleased]: https://github.com/greglook/blocks-s3/compare/0.4.0...HEAD
+[Unreleased]: https://github.com/greglook/blocks-s3/compare/2.0.0...HEAD
+[2.0.0]: https://github.com/greglook/blocks-s3/compare/0.4.0...2.0.0
 [0.4.0]: https://github.com/greglook/blocks-s3/compare/0.3.3...0.4.0
 [0.3.3]: https://github.com/greglook/blocks-s3/compare/0.3.2...0.3.3
 [0.3.2]: https://github.com/greglook/blocks-s3/compare/0.3.1...0.3.2